requirements.txt
# Ascli requirements

-i https://pypi.org/simple
click>=8.0.4; python_version >= '3.6'
colorama>=0.4.4
pillow>=9.0.1
typer>=0.4.0
